remote_repositories/API service/pyproject.toml
[tool.poetry]
name = "api service"
version = "0.1.0"
description = ""
authors = ["Sim Roman <r.sim@vk.team>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
langchain = "^0.0.311"
pdfminer-six = "^20221105"
sentence-transformers = "^2.2.2"
faiss-cpu = "^1.7.4"
tqdm = "^4.66.1"
pandas = "1.5.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
